Shorten state machine verbiage to fsm in drivers to match other places
